requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
mcp>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.8.0
httpx>=0.27.0
pandas>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
coverage>=7.6.0
mypy>=1.11.0
ruff>=0.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
